refactor(mb-api): valeur en Float, fixtures cascadées, helpers ids randoms
- valeur passe en Float (DOUBLE PRECISION) — drop notion de précision Decimal
- fixtures: deps cascadées avec overrides imbriqués (referentielIndividu/relation/valeurAvancement)
- fixtures: update = overrides explicites uniquement (n'écrase plus avec les defaults)
- fixtures: drop surcharge 0-arg sur les helpers cascadés (deps requises)
- helpers testIndicateurId/testDeptId/testRegId (+ testXxxIds(n) tuple typé) pour des publicIds randoms évitant la contention inter-tests
- adapter tous les tests d'intégration mb-api
- ESLint config: désactiver no-unsafe-assignment dans les *.test.ts (expect.any(String))

+++ apps/mb-api/prisma/migrations/20260505161720_add_referentiels_individus_valeurs/migration.sql
@@ -45,7 +45,7 @@ CREATE TABLE "valeur_avancement" (
     "indicateur_id" UUID NOT NULL,
     "individu_id" UUID NOT NULL,
     "date" TEXT NOT NULL,
-    "valeur" DECIMAL(18,6) NOT NULL,
+    "valeur" DOUBLE PRECISION NOT NULL,
     "created_at" TIMESTAMP(3) NOT NULL DEFAULT CURRENT_TIMESTAMP,
     "updated_at" TIMESTAMP(3) NOT NULL,
